// web/src/shared/api/generated/model/keysetPageOfMemberResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Cadence API
 * Project and work management for teams.
 * OpenAPI spec version: 1
 */
import type { MemberResponse } from './memberResponse';

export interface KeysetPageOfMemberResponse {
  items: MemberResponse[];
  /** @nullable */
  nextCursor: string | null;
}